directory refactor and reuse, manual date filtering
