+++ tap_linkedin_ads/streams/ad_analytics/__init__.py
@@ -1,0 +1,1 @@
+"""Tap for LinkedInAds."""
